Update comment about deleting scroll contexts
@jsf9k wrote: These are expensive resources, so there is an upper
limit on the number of them that can be open at once.  They
should be deleted as soon as possible after use, and their
deletion does not actually delete any data from the database.

The comment above was originally written by @jsf9k in PR #2, but
since I needed the "es:ESHttpDelete" permission in order for
the re-skeletonization (PR #3) to be completed properly, #2 is
now obsolete and can be closed.

Co-authored-by: Shane Frasier <jeremy.frasier@gwe.cisa.dhs.gov>